refactor: extract cwl text generation method
This extracts the text generation method from
the log registry class.

Additionally any methods that were now irrelevant
were removed.

IMPORTANT: This commit cannot compile. It will be
resolved in a following commit.
